Too many changes
All the inbox structure is now functional. You can send and receive messages, mark them as read and watch the amount you haven't read IRT.
It only lasts to complete the serach-bar to filter the messages, but for the rest I think is a gret job for just a weekend, isn't it?